requirements.txt
Flask
Flask-Login
psycopg2-binary
reportlab
requests
gunicorn
werkzeug
python-dotenv